requirements.txt
Django>=2.1.3,<3.0.3
djangorestframework>=3.9.0,<3.11.0
psycopg2>=2.7.5,<2.8.4
flake8>=3.6.0,<3.7.9
